backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-genai>=0.1.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
